Added example for a search-suggestions component
This adds an example for a search-suggestions component which
will search through the titles of the articles which are listed
out on the index page of the site. At this point in time there are
a lot of unanswered questions, which I have attempted to formulate
on the `/search-suggestions/index.html` page. Over time and after
testing, I hope that there will be some answers to the questions,
at which time I will update the test to reflect those answers.